refactor: extract package store
